migrate to .net 6
