+++ AssemblyPublicizer.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27703.2042
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AssemblyPublicizer", "AssemblyPublicizer\AssemblyPublicizer.csproj", "{82206A7B-CAA3-4FF9-BBBD-D88B3AAA7BA4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AssemblyPublicizer", "AssemblyPublicizer\AssemblyPublicizer.csproj", "{F8F55ABE-9B9F-45AC-93EC-2C2CAE41A075}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{82206A7B-CAA3-4FF9-BBBD-D88B3AAA7BA4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{82206A7B-CAA3-4FF9-BBBD-D88B3AAA7BA4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{82206A7B-CAA3-4FF9-BBBD-D88B3AAA7BA4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{82206A7B-CAA3-4FF9-BBBD-D88B3AAA7BA4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F8F55ABE-9B9F-45AC-93EC-2C2CAE41A075}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F8F55ABE-9B9F-45AC-93EC-2C2CAE41A075}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F8F55ABE-9B9F-45AC-93EC-2C2CAE41A075}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F8F55ABE-9B9F-45AC-93EC-2C2CAE41A075}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
