--- AssemblyPublicizer/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Mono.Cecil" version="0.10.0" targetFramework="net46" />
-  <package id="Mono.Options" version="5.3.0.1" targetFramework="net46" />
-</packages>
