+++ Licenses/Mono.Cecil.LICENSE.txt
